+++ react-frontend/src/GameLogic/GameTypes.ts
