+++ react-frontend/src/helpers/gameHelpers.ts
@@ -1,4 +1,4 @@
-import { GameStateProps } from "../Game/GameTypes";
+import { GameStateProps } from "../GameLogic/GameTypes";
 import { gameStateSchema, scoreSchema } from "../Schemas/yupSchemas";
 
 export const saveState = async (gameState: GameStateProps) => {
